feat(legal): allow explicit zeroing of upper bits
ZExt will be optimized later on, and we will already deal with i32s.
However, in some cases, we will have to make sure the upper bits are
clean (e.g., when using div, as they play a role in a simulated i8
division).

```
define i8 @udiv(i8 %a, i8 %b) {
entry:
    %0 = udiv i8 %a, %b
    ret i8 %0
}
```

is getting legalized to:
```
udiv:
        i32.const       255
        local.set       2
        local.get       1
        local.get       2
        i32.and
        local.set       3
        local.get       0
        local.get       2
        i32.and
        local.get       3
        i32.div_u
        return
        end_function
```
by LLVM.