fix(binary): assign local index when writing to binary
Each local of a type must be consecutively encoded, so we cannot
assign local values before knowing how many we have. This leads to
locals needing an indirection over a pointer, so we can refer to a
concrete local without assigning it a fixed value.

An alternative would be to fix up locals when writing the binary.